+++ src/instagram/api/types/friendships.nim
@@ -2,6 +2,10 @@ from std/tables import Table
 
 type
   IgFriendshipStatusResponse* = ref object
+    friendshipStatus*: IgFriendshipStatus
+    previousFollowing*: bool
+    status*: string
+  IgFriendshipStatusesResponse* = ref object
     friendshipStatuses*: IgFriendshipStatuses
     status*: string
   IgFriendshipStatuses* = Table[string, IgFriendshipStatus]
@@ -13,3 +17,8 @@ type
     isRestricted*: bool
     outgoingRequest*: bool
     isFeedFavorite*: bool
+    followedBy*: bool
+    blocking*: bool
+    muting*: bool
+    subscribed*: bool
+    isEligibleToSubscribe*: bool